use original addNPC and addPC functions
This calls the original addNPC and addPC functions which should make this less time-consuming to keep updated when the ruleset changes and also improve compatibility with other extensions.
The only downside I see in doing this to addNPC is that pinned effects modifying initiative will not be used for the automatic initiative roll when adding the NPC.
I have not tested these changes very much as I also do not use 5E, but I made these changes to my Pathfinder port and it seems to be working great. I did open 5E with these changes and just drag a NPC and a PC into the combat tracker and they copied in without script errors and with their advanced effects.